+++ example/lib/directed_graphview.dart
@@ -12,38 +12,39 @@ class _GraphClusterViewPageState extends State<GraphClusterViewPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+        appBar: AppBar(),
         body: Column(
-      children: [
-        Expanded(
-          child: InteractiveViewer(
-              constrained: false,
-              boundaryMargin: EdgeInsets.all(8),
-              minScale: 0.001,
-              maxScale: 100,
-              child: GraphView(
-                  graph: graph,
-                  algorithm: builder,
-                  paint: Paint()
-                    ..color = Colors.green
-                    ..strokeWidth = 1
-                    ..style = PaintingStyle.fill,
-                  builder: (Node node) {
-                    // I can decide what widget should be shown here based on the id
-                    var a = node.key.value as int;
-                    if (a == 2) {
-                      return rectangWidget(a);
-                    }
-                    return rectangWidget(a);
-                  })),
-        ),
-      ],
-    ));
+          children: [
+            Expanded(
+              child: InteractiveViewer(
+                  constrained: false,
+                  boundaryMargin: EdgeInsets.all(8),
+                  minScale: 0.001,
+                  maxScale: 100,
+                  child: GraphView(
+                      graph: graph,
+                      algorithm: builder,
+                      paint: Paint()
+                        ..color = Colors.green
+                        ..strokeWidth = 1
+                        ..style = PaintingStyle.fill,
+                      builder: (Node node) {
+                        // I can decide what widget should be shown here based on the id
+                        var a = node.key!.value as int?;
+                        if (a == 2) {
+                          return rectangWidget(a);
+                        }
+                        return rectangWidget(a);
+                      })),
+            ),
+          ],
+        ));
   }
 
   int n = 8;
   Random r = Random();
 
-  Widget rectangWidget(int i) {
+  Widget rectangWidget(int? i) {
     return Container(
         padding: EdgeInsets.all(16),
         decoration: BoxDecoration(
@@ -56,7 +57,7 @@ class _GraphClusterViewPageState extends State<GraphClusterViewPage> {
   }
 
   final Graph graph = Graph();
-  Algorithm builder;
+  late Algorithm builder;
 
   @override
   void initState() {
@@ -65,10 +66,9 @@ class _GraphClusterViewPageState extends State<GraphClusterViewPage> {
     final c = Node.Id(3);
     final d = Node.Id(4);
     final e = Node.Id(5);
-    final f = Node(rectangWidget(
-        6)); //using deprecated mechanism of directly placing the widget here
-    final g = Node(rectangWidget(7));
-    final h = Node(rectangWidget(8));
+    final f = Node.Id(6);
+    final g = Node.Id(7);
+    final h = Node.Id(8);
 
     graph.addEdge(a, b, paint: Paint()..color = Colors.red);
     graph.addEdge(a, c);