+++ example/lib/graph_cluster_animated.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:graphview/GraphView.dart';
@@ -6,7 +7,8 @@ import 'package:graphview/GraphView.dart';
 class GraphScreen extends StatefulWidget {
   Graph graph;
   Algorithm algorithm;
-  final Paint paint;
+  final Paint? paint;
+
   GraphScreen(this.graph, this.algorithm, this.paint);
 
   @override
@@ -15,16 +17,28 @@ class GraphScreen extends StatefulWidget {
 
 class _GraphScreenState extends State<GraphScreen> {
   bool animated = true;
+  Random r = Random();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        automaticallyImplyLeading: true,
         title: Text("Graph Screen"),
         actions: [
           IconButton(
             icon: Icon(Icons.add),
+            onPressed: () async {
+              setState(() {
+                final node12 = Node.Id(r.nextInt(100).toString());
+                var edge = widget.graph.getNodeAtPosition(r.nextInt(widget.graph.nodeCount()));
+                print(edge);
+                widget.graph.addEdge(edge, node12);
+                setState(() {});
+              });
+            },
+          ),
+          IconButton(
+            icon: Icon(Icons.animation),
             onPressed: () async {
               setState(() {
                 animated = !animated;
@@ -41,8 +55,26 @@ class _GraphScreenState extends State<GraphScreen> {
           child: GraphView(
             graph: widget.graph,
             algorithm: widget.algorithm,
+            animated: animated,
+            builder: (Node node) {
+              // I can decide what widget should be shown here based on the id
+              var a = node.key!.value as String;
+              return rectangWidget(a);
+            },
           )),
     );
+  }
+
+  Widget rectangWidget(String? i) {
+    return Container(
+        padding: EdgeInsets.all(16),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(4),
+          boxShadow: [
+            BoxShadow(color: Colors.blue, spreadRadius: 1),
+          ],
+        ),
+        child: Center(child: Text("Node $i")));
   }
 
   Future<void> update() async {