+++ lib/GraphView.dart
@@ -10,18 +10,31 @@ import 'package:flutter/rendering.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 
 part 'Graph.dart';
+
 part 'Algorithm.dart';
+
 part 'edgerenderer/ArrowEdgeRenderer.dart';
+
 part 'edgerenderer/EdgeRenderer.dart';
+
 part 'forcedirected/FruchtermanReingoldAlgorithm.dart';
+
 part 'layered/SugiyamaAlgorithm.dart';
+
 part 'layered/SugiyamaConfiguration.dart';
+
 part 'layered/SugiyamaEdgeData.dart';
+
 part 'layered/SugiyamaEdgeRenderer.dart';
+
 part 'layered/SugiyamaNodeData.dart';
+
 part 'tree/BuchheimWalkerAlgorithm.dart';
+
 part 'tree/BuchheimWalkerConfiguration.dart';
+
 part 'tree/BuchheimWalkerNodeData.dart';
+
 part 'tree/TreeEdgeRenderer.dart';
 
 typedef NodeWidgetBuilder = Widget Function(Node node);
@@ -30,19 +43,12 @@ class GraphView extends StatefulWidget {
   final Graph graph;
   final Algorithm algorithm;
   final Paint? paint;
-  final NodeWidgetBuilder? builder;
-  final bool animated =
-      false; // A later feature, had to include here to migrate to null safety
+  final NodeWidgetBuilder builder;
+  final bool animated;
 
   GraphView(
-      {Key? key,
-      required this.graph,
-      required this.algorithm,
-      this.paint,
-      this.builder})
-      : assert(graph != null),
-        assert(algorithm != null),
-        super(key: key);
+      {Key? key, required this.graph, required this.algorithm, this.paint, required this.builder, this.animated = true})
+      : super(key: key);
 
   @override
   _GraphViewState createState() => _GraphViewState();
@@ -51,8 +57,8 @@ class GraphView extends StatefulWidget {
 class _GraphViewState extends State<GraphView> {
   @override
   Widget build(BuildContext context) {
-    if (widget.animated) {
-      return GraphAnimated(
+    if (widget.algorithm is FruchtermanReingoldAlgorithm) {
+      return _GraphViewAnimated(
         key: widget.key,
         graph: widget.graph,
         algorithm: widget.algorithm,
@@ -76,29 +82,30 @@ class _GraphView extends MultiChildRenderObjectWidget {
   final Algorithm algorithm;
   final Paint? paint;
 
-  _GraphView(
-      {Key? key,
-      required this.graph,
-      required this.algorithm,
-      this.paint,
-      NodeWidgetBuilder? builder})
-      : assert(graph != null),
-        assert(algorithm != null),
-        super(key: key, children: _extractChildren(graph, builder)) {
+  _GraphView({Key? key, required this.graph, required this.algorithm, this.paint, required NodeWidgetBuilder builder})
+      : super(key: key, children: _extractChildren(graph, builder)) {
     assert(() {
+      if (children.isEmpty) {
+        throw FlutterError(
+          'Children must not be empty, ensure you are overriding the builder',
+        );
+      }
+
       return true;
     }());
   }
 
-  // Traverses the InlineSpan tree and depth-first collects the list of
-  // child widgets that are created in WidgetSpans.
-  static List<Widget> _extractChildren(
-      Graph graph, NodeWidgetBuilder? builder) {
+  // Traverses the nodes depth-first collects the list of child widgets that are created.
+  static List<Widget> _extractChildren(Graph graph, NodeWidgetBuilder builder) {
     final result = <Widget>[];
 
     graph.nodes.forEach((node) {
-      result.add(node.data ?? builder!(node));
+      var widget = node.data ?? builder(node);
+      if (widget != null) {
+        result.add(widget);
+      }
     });
+
     return result;
   }
 
@@ -108,8 +115,7 @@ class _GraphView extends MultiChildRenderObjectWidget {
   }
 
   @override
-  void updateRenderObject(
-      BuildContext context, RenderCustomLayoutBox renderObject) {
+  void updateRenderObject(BuildContext context, RenderCustomLayoutBox renderObject) {
     renderObject
       ..graph = graph
       ..algorithm = algorithm
@@ -118,9 +124,7 @@ class _GraphView extends MultiChildRenderObjectWidget {
 }
 
 class RenderCustomLayoutBox extends RenderBox
-    with
-        ContainerRenderObjectMixin<RenderBox, NodeBoxData>,
-        RenderBoxContainerDefaultsMixin<RenderBox, NodeBoxData> {
+    with ContainerRenderObjectMixin<RenderBox, NodeBoxData>, RenderBoxContainerDefaultsMixin<RenderBox, NodeBoxData> {
   late Graph _graph;
   late Algorithm _algorithm;
   late Paint _paint;
@@ -233,28 +237,25 @@ class RenderCustomLayoutBox extends RenderBox
 
 class NodeBoxData extends ContainerBoxParentData<RenderBox> {}
 
-class GraphAnimated extends StatefulWidget {
+class _GraphViewAnimated extends StatefulWidget {
   final Graph graph;
   final Algorithm algorithm;
   final Paint? paint;
-  final result = <Widget>[];
+  final nodes = <Widget>[];
+  final stepMilis = 25;
 
-  GraphAnimated(
-      {Key? key,
-      required this.graph,
-      required this.algorithm,
-      this.paint,
-      NodeWidgetBuilder? builder}) {
+  _GraphViewAnimated(
+      {Key? key, required this.graph, required this.algorithm, this.paint, required NodeWidgetBuilder builder}) {
     graph.nodes.forEach((node) {
-      result.add(node.data ?? builder!(node));
+      nodes.add(node.data ?? builder(node));
     });
   }
 
   @override
-  _GraphAnimatedState createState() => _GraphAnimatedState();
+  _GraphViewAnimatedState createState() => _GraphViewAnimatedState();
 }
 
-class _GraphAnimatedState extends State<GraphAnimated> {
+class _GraphViewAnimatedState extends State<_GraphViewAnimated> {
   late Timer timer;
   late Graph graph;
   late Algorithm algorithm;
@@ -271,7 +272,7 @@ class _GraphAnimatedState extends State<GraphAnimated> {
   }
 
   void startTimer() {
-    timer = Timer.periodic(Duration(milliseconds: 25), (timer) {
+    timer = Timer.periodic(Duration(milliseconds: widget.stepMilis), (timer) {
       algorithm.step(graph);
       update();
     });
@@ -285,8 +286,7 @@ class _GraphAnimatedState extends State<GraphAnimated> {
 
   @override
   Widget build(BuildContext context) {
-    algorithm.setDimensions(
-        MediaQuery.of(context).size.width, MediaQuery.of(context).size.height);
+    algorithm.setDimensions(MediaQuery.of(context).size.width, MediaQuery.of(context).size.height);
 
     return Stack(
       clipBehavior: Clip.none,
@@ -298,7 +298,7 @@ class _GraphAnimatedState extends State<GraphAnimated> {
         ...List<Widget>.generate(graph.nodeCount(), (index) {
           return Positioned(
             child: GestureDetector(
-              child: widget.result[index],
+              child: widget.nodes[index],
               onPanUpdate: (details) {
                 graph.getNodeAtPosition(index).position += details.delta;
                 update();
@@ -321,6 +321,7 @@ class EdgeRender extends CustomPainter {
   Algorithm algorithm;
   Graph graph;
   Offset offset;
+
   EdgeRender(this.algorithm, this.graph, this.offset);
 
   @override