+++ lib/layered/SugiyamaEdgeRenderer.dart
@@ -5,6 +5,7 @@ class SugiyamaEdgeRenderer extends ArrowEdgeRenderer {
   Map<Edge, SugiyamaEdgeData> edgeData;
 
   SugiyamaEdgeRenderer(this.nodeData, this.edgeData);
+
   var path = Path();
 
   void render(Canvas canvas, Graph graph, Paint paint) {
@@ -42,20 +43,14 @@ class SugiyamaEdgeRenderer extends ArrowEdgeRenderer {
         final size = bendPoints.length;
 
         if (nodeData[source]!.isReversed) {
-          clippedLine = clipLine(bendPoints[2], bendPoints[3], bendPoints[0],
-              bendPoints[1], destination);
+          clippedLine = clipLine(bendPoints[2], bendPoints[3], bendPoints[0], bendPoints[1], destination);
         } else {
-          clippedLine = clipLine(bendPoints[size - 4], bendPoints[size - 3],
-              bendPoints[size - 2], bendPoints[size - 1], destination);
+          clippedLine = clipLine(
+              bendPoints[size - 4], bendPoints[size - 3], bendPoints[size - 2], bendPoints[size - 1], destination);
         }
 
         final triangleCentroid = drawTriangle(
-            canvas,
-            edgeTrianglePaint ?? trianglePaint,
-            clippedLine[0],
-            clippedLine[1],
-            clippedLine[2],
-            clippedLine[3]);
+            canvas, edgeTrianglePaint ?? trianglePaint, clippedLine[0], clippedLine[1], clippedLine[2], clippedLine[3]);
 
         path.reset();
         path.moveTo(bendPoints[0], bendPoints[1]);
@@ -75,15 +70,10 @@ class SugiyamaEdgeRenderer extends ArrowEdgeRenderer {
         clippedLine = clipLine(startX, startY, stopX, stopY, destination);
 
         final triangleCentroid = drawTriangle(
-            canvas,
-            edgeTrianglePaint ?? trianglePaint,
-            clippedLine[0],
-            clippedLine[1],
-            clippedLine[2],
-            clippedLine[3]);
+            canvas, edgeTrianglePaint ?? trianglePaint, clippedLine[0], clippedLine[1], clippedLine[2], clippedLine[3]);
 
-        canvas.drawLine(Offset(clippedLine[0], clippedLine[1]),
-            Offset(triangleCentroid[0], triangleCentroid[1]), currentPaint);
+        canvas.drawLine(
+            Offset(clippedLine[0], clippedLine[1]), Offset(triangleCentroid[0], triangleCentroid[1]), currentPaint);
       }
     });
   }